backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.0.1
langchain==0.3.14
langchain-groq==0.2.4
langchain-community==0.3.14
langchain-huggingface==0.1.2
faiss-cpu==1.9.0
sentence-transformers==3.3.1
pypdf2==3.0.1
pypdf==5.1.0
psycopg2-binary==2.9.10
pillow==11.1.0
pytesseract==0.3.13
